+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking:** PDF reports are not generated anymore by default.  Add
   `generate_report = "every_iteration"` in the settings to restore behaviour of previous
   versions.
+- **Breaking:** `Optimizer.save_pdf_report()` is moved out of the `Optimizer` class to
+  `report.produce_optimization_report()`.
 - *Relevant for Dev's only:* Use ruff instead of flake8 for linting.
 
 ### Added